+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Remove NeoForgeModLoader. Mods no longer need to register themselves with
   Ash API on NeoForge
 - Remove empty NeoForge entrypoint. NeoForge does not require one anymore
+- Remove Ash ModInitializers completely
+  1. Mods no longer have to register themselves with Ash API on NeoForge
+  2. NeoForge now allows for client-only entry points
 
 ## v20.6.2
 ### v20.6.2-beta